refactor(memory): 将 main.py 解耦为模块化结构并分离核心组件
