+++ plugin.py
@@ -1,0 +1,979 @@
+"""AstrBot Memora Connect 插件"""
+import asyncio
+import os
+from typing import Any, Dict, List, Optional
+
+from astrbot.api.provider import ProviderRequest
+from astrbot.api.event import filter, AstrMessageEvent, MessageEventResult, MessageChain
+from astrbot.api.star import Context, Star, StarTools
+from astrbot.api import logger
+from astrbot.api import AstrBotConfig
+
+from .memory_system import MemorySystem
+from .memory_graph import MemoryGraph
+from .enhanced_memory_display import EnhancedMemoryDisplay
+from .enhanced_memory_recall import EnhancedMemoryRecall
+from .memory_graph_visualization import MemoryGraphVisualizer
+from .web_server import MemoryWebServer
+from .memory_events import MemoryEventBus, MemoryEvent, MemoryEventType, initialize_event_bus, shutdown_event_bus
+from .topic_engine import TopicEngine
+from .user_profiling import UserProfilingSystem
+from .temporal_memory import TemporalMemorySystem
+from .memory_api_gateway import MemoryAPIGateway
+from .resource_management import resource_manager
+
+
+class MemoraConnectPlugin(Star):
+    def __init__(self, context: Context, config: AstrBotConfig):
+        super().__init__(context)
+        data_dir = StarTools.get_data_dir() / "memora_connect"
+        self.memory_system = MemorySystem(context, config, data_dir)
+        self.memory_display = EnhancedMemoryDisplay(self.memory_system)
+        self.graph_visualizer = MemoryGraphVisualizer(self.memory_system)
+        self._initialized = False
+        self.web_server = None
+        
+        # 新增：主动能力升级模块
+        self.event_bus = None
+        self.topic_engine = None
+        self.user_profiling = None
+        self.temporal_memory = None
+        self.api_gateway = None
+        
+        asyncio.create_task(self._async_init())
+    
+    def _debug_log(self, message: str, level: str = "debug"):
+        try:
+            if level == "debug":
+                logger.debug(message)
+            elif level == "info":
+                logger.info(message)
+            elif level == "warning":
+                logger.warning(message)
+            elif level == "error":
+                logger.error(message)
+            else:
+                logger.info(message)
+        except Exception:
+            pass
+    
+    async def _async_init(self):
+        """异步初始化包装器"""
+        try:
+            logger.info("开始异步初始化记忆系统...")
+            await self.memory_system.initialize()
+            
+            # 初始化新增模块
+            try:
+                logger.info("初始化主动能力升级模块...")
+                
+                # 1. 初始化事件总线
+                self.event_bus = await initialize_event_bus()
+                logger.info("✓ 事件总线已启动")
+                
+                # 2. 初始化话题引擎
+                self.topic_engine = TopicEngine(self.memory_system)
+                logger.info("✓ 话题引擎已初始化")
+                
+                # 3. 初始化用户画像系统
+                self.user_profiling = UserProfilingSystem(self.memory_system)
+                logger.info("✓ 用户画像系统已初始化")
+                
+                # 4. 初始化时间维度记忆系统
+                self.temporal_memory = TemporalMemorySystem(self.memory_system)
+                logger.info("✓ 时间维度记忆系统已初始化")
+                
+                # 5. 初始化API网关
+                self.api_gateway = MemoryAPIGateway(
+                    self.memory_system,
+                    self.topic_engine,
+                    self.user_profiling,
+                    self.temporal_memory
+                )
+                logger.info("✓ API网关已初始化")
+                
+                logger.info("主动能力升级模块初始化完成！")
+            except Exception as upgrade_e:
+                logger.error(f"主动能力升级模块初始化失败: {upgrade_e}", exc_info=True)
+            
+            self._initialized = True
+            
+            # 根据配置启动 Web 界面
+            try:
+                web_cfg = (self.memory_system.memory_config or {}).get("web_ui", {}) or {}
+                if web_cfg.get("enabled", False):
+                    host = str(web_cfg.get("host", "127.0.0.1"))
+                    port = int(web_cfg.get("port", 8350))
+                    token = str(web_cfg.get("access_token", "") or "")
+                    self.web_server = MemoryWebServer(self.memory_system, host=host, port=port, access_token=token)
+                    await self.web_server.start()
+                    logger.info(f"Web 界面已启动: http://{host}:{port}")
+            except Exception as _we:
+                logger.error(f"启动Web界面失败: {_we}", exc_info=True)
+            logger.info("记忆系统异步初始化完成")
+        except Exception as e:
+            logger.error(f"记忆系统初始化失败: {e}", exc_info=True)
+        
+    @filter.command_group("记忆")
+    def memory(self):
+        """记忆管理指令组"""
+        pass
+
+    @memory.command("回忆")
+    async def memory_recall(self, event: AstrMessageEvent, keyword: str):
+        # 检查记忆系统是否启用
+        if not self.memory_system.config_manager.is_memory_system_enabled():
+            yield event.plain_result("记忆系统已禁用，无法使用回忆功能。")
+            return
+        memories = await self.memory_system.recall_memories_full(keyword)
+        response = self.memory_display.format_memory_search_result(memories, keyword)
+        yield event.plain_result(response)
+
+    @memory.command("状态")
+    async def memory_status(self, event: AstrMessageEvent):
+        # 检查记忆系统是否启用
+        if not self.memory_system.config_manager.is_memory_system_enabled():
+            yield event.plain_result("记忆系统已禁用，无法查看状态。")
+            return
+            
+        stats = self.memory_display.format_memory_statistics()
+        yield event.plain_result(stats)
+    @memory.command("印象")
+    async def memory_impression(self, event: AstrMessageEvent, name: str):
+        """查询人物印象摘要和相关记忆"""
+        # 检查记忆系统是否启用
+        if not self.memory_system.config_manager.is_memory_system_enabled():
+            yield event.plain_result("记忆系统已禁用，无法查询印象。")
+            return
+            
+        try:
+            # 获取群组ID
+            group_id = self.memory_system._extract_group_id_from_event(event)
+            
+            # 获取印象摘要
+            impression_summary = self.memory_system.get_person_impression_summary(group_id, name)
+            
+            # 获取印象记忆列表
+            impression_memories = self.memory_system.get_person_impression_memories(group_id, name, limit=5)
+            
+            # 格式化响应
+            response_parts = []
+            
+            # 添加印象摘要
+            if impression_summary:
+                score = impression_summary.get("score", 0.5)
+                score_desc = self.memory_system._score_to_description(score)
+                response_parts.append(f"📝 {name} 的印象摘要:")
+                response_parts.append(f"   印象: {impression_summary.get('summary', '无')}")
+                response_parts.append(f"   好感度: {score_desc} ({score:.2f})")
+                response_parts.append(f"   记忆数: {impression_summary.get('memory_count', 0)}")
+                response_parts.append(f"   更新时间: {impression_summary.get('last_updated', '无')}")
+            else:
+                response_parts.append(f"📝 尚未建立对 {name} 的印象")
+            
+            # 添加相关记忆
+            if impression_memories:
+                response_parts.append("\n📚 相关记忆:")
+                for i, memory in enumerate(impression_memories, 1):
+                    response_parts.append(f"   {i}. {memory['content']}")
+                    if memory.get('details'):
+                        response_parts.append(f"      详情: {memory['details']}")
+                    response_parts.append(f"      好感度: {memory['score']:.2f} | 时间: {memory['last_accessed']}")
+            else:
+                response_parts.append(f"\n📚 暂无关于 {name} 的印象记忆")
+            
+            # 组合响应
+            response = "\n".join(response_parts)
+            yield event.plain_result(response)
+            
+        except Exception as e:
+            logger.error(f"查询印象失败: {e}")
+            yield event.plain_result(f"查询 {name} 的印象时出现错误")
+    
+    @memory.command("图谱")
+    async def memory_graph(self, event: AstrMessageEvent, layout_style: str = "auto"):
+        """生成记忆图谱可视化图片
+        
+        Args:
+            layout_style: 布局风格，可选值：
+                - auto: 自适应布局（根据图的复杂度自动选择最适合的布局，默认）
+                - force_directed: 力导向布局
+                - circular: 圆形布局
+                - kamada_kawai: Kamada-Kawai布局
+                - spectral: 谱布局
+                - community: 社区布局
+                - hierarchical: 多层次布局
+        """
+        # 检查记忆系统是否启用
+        if not self.memory_system.config_manager.is_memory_system_enabled():
+            yield event.plain_result("记忆系统已禁用，无法生成图谱。")
+            return
+            
+        try:
+            # 发送生成中的提示
+            yield event.plain_result(f"🔄 正在生成记忆图谱（布局风格: {layout_style}），请稍候...")
+            
+            # 异步生成图谱图片
+            image_path = await self.graph_visualizer.generate_graph_image(layout_style=layout_style)
+            
+            if image_path:
+                # 检查文件是否存在
+                if os.path.exists(image_path):
+                    # 发送图片消息
+                    try:
+                        # 尝试使用 AstrBot 的图片发送功能
+                        if hasattr(event, 'send_image'):
+                            await event.send_image(image_path)
+                            yield event.plain_result(f"✅ 记忆图谱已生成！（布局风格: {layout_style}）")
+                        else:
+                            # 如果不支持直接发送图片，尝试使用其他方法
+                            yield event.image_result(image_path)
+                    except Exception as img_e:
+                        logger.error(f"发送图片失败: {img_e}", exc_info=True)
+                        # 如果发送图片失败，发送文件路径
+                        yield event.plain_result(f"✅ 记忆图谱已生成！（布局风格: {layout_style}）\n图片路径: {image_path}")
+                else:
+                    yield event.plain_result("❌ 图谱文件生成失败，请检查权限和磁盘空间。")
+            else:
+                yield event.plain_result("❌ 记忆图谱生成失败，可能是因为：\n1. 未安装依赖库（networkx, matplotlib）\n2. 记忆数据为空\n3. 系统错误")
+                
+        except Exception as e:
+            logger.error(f"生成记忆图谱失败: {e}", exc_info=True)
+            yield event.plain_result(f"❌ 生成记忆图谱时出现错误: {str(e)}")
+    
+    @filter.event_message_type(filter.EventMessageType.ALL)
+    async def on_message(self, event: AstrMessageEvent):
+        """监听所有消息，形成记忆并注入相关记忆"""
+        if not self._initialized:
+            self._debug_log("记忆系统尚未初始化完成，跳过消息处理", "debug")
+            return
+        # 检查记忆系统是否启用
+        if not self.memory_system.config_manager.is_memory_system_enabled():
+            return
+            
+        try:
+            # 提取群聊ID，用于群聊隔离
+            group_id = event.get_group_id() if event.get_group_id() else ""
+            
+            # 1. 为当前群聊加载相应的记忆状态（异步优化）
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                # 清空当前记忆图，重新加载群聊特定的记忆
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+            
+            # 2. 注入相关记忆到上下文（快速异步操作）
+            self.memory_system._create_managed_task(self.memory_system.inject_memories_to_context(event))
+            
+            # 3. 消息处理使用异步队列，避免阻塞主流程
+            self.memory_system._create_managed_task(self._process_message_async(event, group_id))
+                
+        except Exception as e:
+            self._debug_log(f"on_message处理错误: {e}", "error")
+    
+    async def _process_message_async(self, event: AstrMessageEvent, group_id: str):
+        """异步消息处理，避免阻塞主流程"""
+        try:
+            message = event.message_str
+            sender_id = event.get_sender_id()
+            
+            # 使用优化后的单次LLM调用处理消息
+            await self.memory_system.process_message_optimized(event, group_id)
+            
+            # === 新增：主动能力升级相关处理 ===
+            if self.topic_engine and self.user_profiling and self.temporal_memory:
+                try:
+                    # 1. 话题追踪
+                    await self.topic_engine.add_message_to_topic(message, sender_id, group_id)
+                    
+                    # 2. 未闭合话题检测
+                    await self.temporal_memory.auto_detect_and_track_questions(message, sender_id, group_id)
+                    
+                    # 3. 禁忌词自动学习
+                    await self.user_profiling.learn_taboo_from_message(sender_id, message, group_id)
+                    
+                    # 4. 查找复活的话题
+                    resurrected = await self.topic_engine.find_resurrected_topics(message, group_id, silence_days=7)
+                    if resurrected:
+                        logger.info(f"检测到复活话题: {resurrected}")
+                    
+                except Exception as upgrade_e:
+                    logger.debug(f"主动能力升级处理失败: {upgrade_e}")
+            
+            # 使用队列化保存，减少I/O操作
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                await self.memory_system._queue_save_memory_state(group_id)
+            else:
+                await self.memory_system._queue_save_memory_state("")  # 默认数据库
+                
+        except Exception as e:
+            self._debug_log(f"异步消息处理失败: {e}", "error")
+
+    @filter.on_llm_request()
+    async def on_llm_request(self, event: AstrMessageEvent, req: ProviderRequest):
+        """处理LLM请求时的记忆召回"""
+        try:
+            if not self._initialized:
+                return
+                
+            # 获取当前消息内容
+            current_message = event.message_str.strip()
+            if not current_message:
+                return
+            
+            # 使用增强记忆召回系统
+            enhanced_recall = EnhancedMemoryRecall(self.memory_system)
+            results = await enhanced_recall.recall_all_relevant_memories(
+                query=current_message,
+                max_memories=self.memory_system.memory_config.get("max_injected_memories", 5)
+            )
+            
+            if results:
+                # 格式化记忆为上下文
+                memory_context = enhanced_recall.format_memories_for_llm(results)
+                
+                # 将记忆注入到系统提示中
+                if hasattr(req, 'system_prompt'):
+                    original_prompt = req.system_prompt or ""
+                    if "【相关记忆】" not in original_prompt:
+                        req.system_prompt = f"{original_prompt}\n\n{memory_context}"
+                        logger.debug(f"已注入 {len(results)} 条记忆到LLM上下文")
+                        
+        except Exception as e:
+            logger.error(f"LLM请求记忆召回失败: {e}", exc_info=True)
+    
+    async def terminate(self):
+        """插件卸载时保存记忆并清理资源"""
+        self._debug_log("开始插件终止流程，清理所有资源", "info")
+        
+        try:
+            # === 新增：清理主动能力升级模块 ===
+            try:
+                if self.event_bus:
+                    logger.info("关闭事件总线...")
+                    await shutdown_event_bus()
+                    logger.info("✓ 事件总线已关闭")
+            except Exception as bus_e:
+                logger.warning(f"关闭事件总线失败: {bus_e}")
+            
+            # 停止 Web 服务
+            if hasattr(self, 'web_server') and self.web_server:
+                try:
+                    await self.web_server.stop()
+                except Exception as _we:
+                    logger.warning(f"停止Web服务失败: {_we}")
+            # 1. 停止维护循环
+            if hasattr(self.memory_system, '_should_stop_maintenance'):
+                self.memory_system._should_stop_maintenance.set()
+            if hasattr(self.memory_system, '_maintenance_task') and self.memory_system._maintenance_task:
+                # 等待维护任务正常退出
+                try:
+                    await asyncio.wait_for(self.memory_system._maintenance_task, timeout=10.0)
+                except asyncio.TimeoutError:
+                    # 如果超时，取消任务
+                    self.memory_system._maintenance_task.cancel()
+                    try:
+                        await self.memory_system._maintenance_task
+                    except asyncio.CancelledError:
+                        pass
+                        
+            # 2. 取消所有托管的异步任务
+            if hasattr(self.memory_system, '_managed_tasks'):
+                await self.memory_system._cancel_all_managed_tasks()
+            
+            # 3. 等待待处理的保存任务完成
+            if hasattr(self.memory_system, '_pending_save_task') and self.memory_system._pending_save_task and not self.memory_system._pending_save_task.done():
+                try:
+                    await asyncio.wait_for(self.memory_system._pending_save_task, timeout=5.0)
+                except asyncio.TimeoutError:
+                    self.memory_system._pending_save_task.cancel()
+                    try:
+                        await self.memory_system._pending_save_task
+                    except asyncio.CancelledError:
+                        pass
+                    
+            # 4. 清理嵌入向量缓存
+            if hasattr(self.memory_system, 'embedding_cache') and self.memory_system.embedding_cache:
+                try:
+                    await self.memory_system.embedding_cache.cleanup()
+                except Exception as e:
+                    logger.warning(f"清理嵌入向量缓存时出错: {e}")
+        
+            # 5. 保存记忆状态
+            await self.memory_system.save_memory_state()
+            
+            # 6. 如果启用了群聊隔离，保存所有群聊数据库
+            if self.memory_system.memory_config.get("enable_group_isolation", True):
+                db_dir = os.path.dirname(self.memory_system.db_path)
+                if os.path.exists(db_dir):
+                    for filename in os.listdir(db_dir):
+                        if filename.startswith("memory_group_") and filename.endswith(".db"):
+                            group_id = filename[12:-3]
+                            await self.memory_system.save_memory_state(group_id)
+            
+            # 7. 使用资源管理器清理所有资源
+            resource_manager.cleanup()
+            
+            self._debug_log("记忆系统已保存并安全关闭", "info")
+            
+        except Exception as e:
+            logger.error(f"插件终止过程中发生错误: {e}", exc_info=True)
+            
+    async def safe_cleanup(self):
+        """安全清理方法，用于在 terminate 之外调用的情况"""
+        await self.terminate()
+
+    # ---------- 插件API ----------
+    async def add_memory_api(self, content: str, theme: str, group_id: str = "", details: str = "", participants: str = "", location: str = "", emotion: str = "", tags: str = "") -> Optional[str]:
+        """
+        【API】添加一条记忆。
+        :param content: 记忆的核心内容。
+        :param theme: 记忆的主题或关键词，用逗号分隔。
+        :param group_id: 群组ID，如果需要在特定群聊中操作。
+        :param details: 记忆的详细信息。
+        :param participants: 参与者，用逗号分隔。
+        :param location: 相关地点。
+        :param emotion: 情感色彩。
+        :param tags: 标签，用逗号分隔。
+        :return: 成功则返回记忆ID，否则返回None。
+        """
+        if not self._initialized or not self.memory_system.memory_system_enabled:
+            logger.warning("API调用失败：记忆系统未启用或未初始化。")
+            return None
+        
+        try:
+            # 切换到正确的群聊上下文
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+
+            concept_id = self.memory_system.memory_graph.add_concept(theme)
+            memory_id = self.memory_system.memory_graph.add_memory(
+                content=content,
+                concept_id=concept_id,
+                details=details,
+                participants=participants,
+                location=location,
+                emotion=emotion,
+                tags=tags,
+                group_id=group_id
+            )
+            
+            # 异步保存
+            await self.memory_system._queue_save_memory_state(group_id)
+            
+            logger.info(f"通过API添加记忆成功: {memory_id}")
+            return memory_id
+        except Exception as e:
+            logger.error(f"API add_memory_api 失败: {e}", exc_info=True)
+            return None
+
+    async def recall_memories_api(self, keyword: str, group_id: str = "") -> List[Dict[str, Any]]:
+        """
+        【API】根据关键词回忆相关记忆。
+        :param keyword: 要查询的关键词。
+        :param group_id: 群组ID，如果需要在特定群聊中操作。
+        :return: 记忆对象字典的列表。
+        """
+        if not self._initialized or not self.memory_system.memory_system_enabled:
+            logger.warning("API调用失败：记忆系统未启用或未初始化。")
+            return []
+
+        try:
+            # 切换到正确的群聊上下文
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+
+            memories = await self.memory_system.recall_memories_full(keyword)
+            return [memory.__dict__ for memory in memories]
+        except Exception as e:
+            logger.error(f"API recall_memories_api 失败: {e}", exc_info=True)
+            return []
+
+    async def record_impression_api(self, person_name: str, summary: str, score: Optional[float], details: str = "", group_id: str = "") -> bool:
+        """
+        【API】记录对某个人的印象。
+        :param person_name: 人物名称。
+        :param summary: 印象摘要。
+        :param score: 好感度分数 (0-1)。
+        :param details: 详细信息。
+        :param group_id: 群组ID。
+        :return: 操作是否成功。
+        """
+        if not self._initialized or not self.memory_system.memory_system_enabled:
+            logger.warning("API调用失败：记忆系统未启用或未初始化。")
+            return False
+
+        try:
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+
+            memory_id = self.memory_system.record_person_impression(group_id, person_name, summary, score, details)
+            await self.memory_system._queue_save_memory_state(group_id)
+            return bool(memory_id)
+        except Exception as e:
+            logger.error(f"API record_impression_api 失败: {e}", exc_info=True)
+            return False
+
+    async def get_impression_summary_api(self, person_name: str, group_id: str = "") -> Optional[Dict[str, Any]]:
+        """
+        【API】获取对某个人的印象摘要。
+        :param person_name: 人物名称。
+        :param group_id: 群组ID。
+        :return: 包含摘要信息的字典，或在找不到时返回None。
+        """
+        if not self._initialized or not self.memory_system.memory_system_enabled:
+            logger.warning("API调用失败：记忆系统未启用或未初始化。")
+            return None
+
+        try:
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+
+            return self.memory_system.get_person_impression_summary(group_id, person_name)
+        except Exception as e:
+            logger.error(f"API get_impression_summary_api 失败: {e}", exc_info=True)
+            return None
+
+    async def adjust_impression_score_api(self, person_name: str, delta: float, group_id: str = "") -> Optional[float]:
+        """
+        【API】调整对某个人的好感度分数。
+        :param person_name: 人物名称。
+        :param delta: 好感度调整量，可正可负。
+        :param group_id: 群组ID。
+        :return: 调整后的新分数，或在失败时返回None。
+        """
+        if not self._initialized or not self.memory_system.memory_system_enabled:
+            logger.warning("API调用失败：记忆系统未启用或未初始化。")
+            return None
+
+        try:
+            if group_id and self.memory_system.memory_config.get("enable_group_isolation", True):
+                self.memory_system.memory_graph = MemoryGraph()
+                self.memory_system.load_memory_state(group_id)
+
+            new_score = self.memory_system.adjust_impression_score(group_id, person_name, delta)
+            await self.memory_system._queue_save_memory_state(group_id)
+            return new_score
+        except Exception as e:
+            logger.error(f"API adjust_impression_score_api 失败: {e}", exc_info=True)
+            return None
+
+    # ---------- 主动能力升级 API ----------
+    async def get_topic_relevance_api(self, message: str, group_id: str = "", max_results: int = 5) -> List[Dict]:
+        """
+        【API】获取消息与现有话题的相关性
+        :param message: 当前消息
+        :param group_id: 群组ID
+        :param max_results: 最多返回结果数
+        :return: [{topic_id, relevance_score, topic_info}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.get_topic_relevance(message, group_id, max_results)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API get_topic_relevance_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def get_intimacy_api(self, user_id: str, group_id: str = "") -> Optional[Dict]:
+        """
+        【API】获取用户亲密度
+        :param user_id: 用户ID
+        :param group_id: 群组ID
+        :return: {score: 0-100, sub_scores: {frequency, depth, emotional}}
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return None
+        
+        try:
+            response = await self.api_gateway.get_intimacy(user_id, group_id)
+            return response.data if response.success else None
+        except Exception as e:
+            logger.error(f"API get_intimacy_api 失败: {e}", exc_info=True)
+            return None
+    
+    async def batch_get_intimacy_api(self, user_ids: List[str], group_id: str = "") -> List[Dict]:
+        """
+        【API】批量获取用户亲密度
+        :param user_ids: 用户ID列表
+        :param group_id: 群组ID
+        :return: [{user_id, score, sub_scores}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.batch_get_intimacy(user_ids, group_id)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API batch_get_intimacy_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def get_user_interests_api(self, user_id: str, group_id: str = "") -> List[Dict]:
+        """
+        【API】获取用户兴趣偏好（TOP 5）
+        :param user_id: 用户ID
+        :param group_id: 群组ID
+        :return: [{concept, weight}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.get_user_interests(user_id, group_id)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API get_user_interests_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def check_taboo_api(self, user_id: str, content: str, group_id: str = "") -> Dict:
+        """
+        【API】检查内容是否包含用户禁忌词
+        :param user_id: 用户ID
+        :param content: 要检查的内容
+        :param group_id: 群组ID
+        :return: {has_taboo: bool, taboo_words: [...]}
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return {"has_taboo": False, "taboo_words": []}
+        
+        try:
+            response = await self.api_gateway.check_taboo(user_id, content, group_id)
+            return response.data if response.success else {"has_taboo": False, "taboo_words": []}
+        except Exception as e:
+            logger.error(f"API check_taboo_api 失败: {e}", exc_info=True)
+            return {"has_taboo": False, "taboo_words": []}
+    
+    async def get_open_topics_api(self, group_id: str = "", days: int = 7) -> List[Dict]:
+        """
+        【API】获取未闭合话题列表
+        :param group_id: 群组ID
+        :param days: 查询最近N天的话题
+        :return: [{topic_id, question, asker_id, days_ago, ...}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.get_open_topics(group_id, days)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API get_open_topics_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def get_today_anniversaries_api(self, group_id: str = "") -> List[Dict]:
+        """
+        【API】获取历史今日记忆
+        :param group_id: 群组ID
+        :return: [{memory_id, event_description, days_ago, ...}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.get_today_anniversaries(group_id)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API get_today_anniversaries_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def find_connection_api(self, user_a: str, user_b: str, group_id: str = "") -> Optional[Dict]:
+        """
+        【API】查找两个用户的关系路径（共同兴趣话题）
+        :param user_a: 用户A的ID
+        :param user_b: 用户B的ID
+        :param group_id: 群组ID
+        :return: {common_topics: [...], connection_strength: float, ...}
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return None
+        
+        try:
+            response = await self.api_gateway.find_connection(user_a, user_b, group_id)
+            return response.data if response.success else None
+        except Exception as e:
+            logger.error(f"API find_connection_api 失败: {e}", exc_info=True)
+            return None
+    
+    async def get_memory_importance_ranking_api(self, group_id: str = "", top_k: int = 10) -> List[Dict]:
+        """
+        【API】获取记忆重要性排序（基于度中心性和激活频率）
+        :param group_id: 群组ID
+        :param top_k: 返回TOP K个记忆
+        :return: [{memory_id, content, importance_score, ...}, ...]
+        """
+        if not self._initialized or not self.api_gateway:
+            logger.warning("API调用失败：系统未初始化")
+            return []
+        
+        try:
+            response = await self.api_gateway.get_memory_importance_ranking(group_id, top_k)
+            return response.data if response.success else []
+        except Exception as e:
+            logger.error(f"API get_memory_importance_ranking_api 失败: {e}", exc_info=True)
+            return []
+    
+    async def subscribe_event_api(self, event_type_str: str, callback) -> bool:
+        """
+        【API】订阅记忆事件
+        :param event_type_str: 事件类型字符串，如 "memory.triggered", "topic.resurrected"
+        :param callback: 回调函数，签名应为 async def callback(event: MemoryEvent)
+        :return: 是否订阅成功
+        """
+        if not self._initialized or not self.event_bus:
+            logger.warning("API调用失败：事件总线未初始化")
+            return False
+        
+        try:
+            # 字符串转事件类型
+            for event_type in MemoryEventType:
+                if event_type.value == event_type_str:
+                    return self.event_bus.subscribe(event_type, callback)
+            
+            logger.warning(f"未知的事件类型: {event_type_str}")
+            return False
+        except Exception as e:
+            logger.error(f"API subscribe_event_api 失败: {e}", exc_info=True)
+            return False
+    
+    async def health_check_api(self) -> Dict:
+        """
+        【API】记忆系统健康检查
+        :return: {healthy: bool, components: {...}, performance: {...}}
+        """
+        if not self._initialized or not self.api_gateway:
+            return {
+                "healthy": False,
+                "error": "系统未初始化"
+            }
+        
+        try:
+            return await self.api_gateway.health_check()
+        except Exception as e:
+            logger.error(f"API health_check_api 失败: {e}", exc_info=True)
+            return {
+                "healthy": False,
+                "error": str(e)
+            }
+
+    # ---------- LLM 函数工具 ----------
+    @filter.llm_tool(name="create_memory")
+    async def create_memory_tool(
+        self,
+        event: AstrMessageEvent,
+        content: str,
+        theme: str = None,
+        topic: str = None,
+        details: str = "",
+        participants: str = "",
+        location: str = "",
+        emotion: str = "",
+        tags: str = "",
+        confidence: str = "0.7"
+    ) -> MessageEventResult:
+        """通过LLM调用获取个记忆点和主题
+
+        Args:
+            content(string): 需要记录的完整对话内容
+            theme(string): 核心关键词，用逗号分隔
+            topic(string): 该记忆所属的主题或关键词（向后兼容）
+            details(string): 具体细节和背景信息
+            participants(string): 涉及的人物，用逗号分隔。特别注意：如果是Bot的发言，请使用"我"作为参与者
+            location(string): 相关场景或地点
+            emotion(string): 情感色彩，如"开心,兴奋"
+            tags(string): 分类标签，如"工作,重要"
+            confidence(number): 置信度，0-1之间的数值
+        """
+        try:
+            # 向后兼容性处理：如果提供了topic但没有theme，使用topic作为theme
+            actual_theme = theme or topic
+            if not actual_theme:
+                logger.warning("创建记忆失败：主题为空")
+                
+                return "创建记忆失败：主题为空"
+            
+            # 参数验证和清理
+            if not content:
+                logger.warning("创建记忆失败：内容为空")
+                
+                return "创建记忆失败：内容为空"
+            
+            # 清理特殊字符
+            import re
+            actual_theme = re.sub(r'[^\w\u4e00-\u9fff,，]', '', str(actual_theme))
+            details = str(details).strip()
+            participants = str(participants).strip()
+            location = str(location).strip()
+            emotion = str(emotion).strip()
+            tags = str(tags).strip()
+            
+            # 将confidence从字符串转换为浮点数
+            try:
+                confidence_float = max(0.0, min(1.0, float(confidence)))
+            except (ValueError, TypeError):
+                logger.warning(f"无法将confidence '{confidence}' 转换为浮点数，使用默认值0.7")
+                confidence_float = 0.7
+            
+            # 创建概念
+            concept_id = self.memory_system.memory_graph.add_concept(actual_theme)
+            
+            # 根据置信度调整记忆强度
+            base_strength = 1.0
+            adjusted_strength = base_strength * confidence_float
+            
+            # 获取群组ID
+            group_id = self.memory_system._extract_group_id_from_event(event)
+            
+            # 创建丰富记忆
+            memory_id = self.memory_system.memory_graph.add_memory(
+                content=content,
+                concept_id=concept_id,
+                details=details,
+                participants=participants,
+                location=location,
+                emotion=emotion,
+                tags=tags,
+                strength=adjusted_strength
+            )
+            
+            logger.info(f"LLM工具创建丰富记忆：{actual_theme} -> {content} (置信度: {confidence})")
+            
+            # 返回空字符串让LLM继续其自然回复流程
+            return f"记忆创建成功,内容为:{content}"
+            
+        except Exception as e:
+            logger.error(f"LLM工具创建记忆失败：{e}")
+            await event.send(MessageChain().message("记忆创建失败"))
+            return "记忆创建失败"
+
+    @filter.llm_tool(name="recall_memory")
+    async def recall_memory_tool(self, event: AstrMessageEvent, keyword: str) -> MessageEventResult:
+        """召回所有相关记忆，包括联想记忆。
+
+        Args:
+            keyword(string): 要查询的关键词或内容
+        """
+        try:
+            enhanced_recall = EnhancedMemoryRecall(self.memory_system)
+            results = await enhanced_recall.recall_all_relevant_memories(
+                query=keyword,
+                max_memories=8
+            )
+            
+            if results:
+                # 生成增强的上下文
+                formatted_memories = enhanced_recall.format_memories_for_llm(results)
+                return f"记忆召回结果:{formatted_memories}"
+            else:
+                # 返回空字符串让LLM继续其自然回复流程
+                return "没有相关记忆"
+                  
+        except Exception as e:
+            logger.error(f"增强记忆召回工具失败：{e}")
+            await event.send(MessageChain().message("记忆召回失败"))
+            return "记忆召回失败"
+
+    @filter.llm_tool(name="adjust_impression")
+    async def adjust_impression_tool(
+        self,
+        event: AstrMessageEvent,
+        person_name: str,
+        delta: str,
+        reason: str = ""
+    ) -> MessageEventResult:
+        """调整对某人的印象和好感度
+
+        Args:
+            person_name(string): 人物名称
+            delta(number): 好感度调整量，可正可负
+            reason(string): 调整原因和详细信息
+        """
+        try:
+            # 获取群组ID
+            group_id = self.memory_system._extract_group_id_from_event(event)
+            
+            # 调整印象分数 - 将字符串转换为浮点数
+            try:
+                delta_float = float(delta)
+            except (ValueError, TypeError):
+                logger.warning(f"无法将delta '{delta}' 转换为浮点数，使用默认值0.0")
+                delta_float = 0.0
+            
+            new_score = self.memory_system.adjust_impression_score(group_id, person_name, delta_float)
+            
+            # 记录调整原因
+            if reason:
+                summary = f"调整对{person_name}的印象：{reason}，当前好感度：{new_score:.2f}"
+                self.memory_system.record_person_impression(group_id, person_name, summary, new_score, reason)
+            
+            logger.info(f"LLM工具调整印象：{person_name} 调整量:{delta} 新分数:{new_score:.2f}")
+            
+            # 返回空字符串让LLM继续其自然回复流程
+            return f"调整印象成功，{person_name} 的好感度为 {new_score:.2f}"
+            
+        except Exception as e:
+            logger.error(f"LLM工具调整印象失败：{e}")
+            await event.send(MessageChain().message("调整印象失败"))
+            return "调整印象失败"
+
+    @filter.llm_tool(name="record_impression")
+    async def record_impression_tool(
+        self,
+        event: AstrMessageEvent,
+        person_name: str,
+        summary: str,
+        score: str = None,
+        details: str = ""
+    ) -> MessageEventResult:
+        """记录或更新对某人的印象
+
+        Args:
+            person_name(string): 人物名称
+            summary(string): 印象摘要描述
+            score(number): 好感度分数 (0-1)，可选
+            details(string): 详细信息和背景
+        """
+        try:
+            # 获取群组ID
+            group_id = self.memory_system._extract_group_id_from_event(event)
+            
+            # 验证分数范围 - 将字符串转换为浮点数
+            score_float = None
+            if score is not None:
+                try:
+                    score_float = max(0.0, min(1.0, float(score)))
+                except (ValueError, TypeError):
+                    logger.warning(f"无法将score '{score}' 转换为浮点数，使用默认值")
+                    score_float = None
+            
+            # 记录印象
+            memory_id = self.memory_system.record_person_impression(
+                group_id, person_name, summary, score_float, details
+            )
+            
+            if memory_id:
+                current_score = self.memory_system.get_impression_score(group_id, person_name)
+                logger.info(f"LLM工具记录印象：{person_name} 分数:{current_score:.2f} 摘要:{summary[:50]}...")
+            
+            # 返回空字符串让LLM继续其自然回复流程
+            return f"记录印象成功，{person_name} 的好感度为 {current_score:.2f}"
+            
+        except Exception as e:
+            logger.error(f"LLM工具记录印象失败：{e}")
+            await event.send(MessageChain().message("记录印象失败"))
+            return "记录印象失败"
+
+
